# swagger/swagger.yaml
swagger: '2.0'
info:
  title: sta_access_manager
  description: sta_access_manager => 入館画面用
  version: 1.0.0
host: hostname
schemes:
- https
- http
basePath: /
produces:
- application/json
paths:
  /register:
    post:
      summary: ユーザー/ICカード登録
      parameters:
      - in: body
        name: user
        description: The user to create.
        schema:
          type: object
          required:
          - id
          - user
          properties:
            id:
              description: The user to create.
              type: string
            user:
              type: object
              required:
              - mail
              - name
              properties:
                mail:
                  type: string
                name:
                  type: string
      tags:
      - User
      responses:
        '201':
          description: 登録成功
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  mail:
                    type: string
                  name:
                    type: string
        '400':
          description: bad request
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  mail:
                    type: string
                  name:
                    type: string
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
  /update:
    put:
      summary: ユーザー情報更新
      parameters:
      - in: body
        name: user
        description: ユーザー情報更新
        schema:
          type: object
          required:
          - id
          - user
          properties:
            id:
              type: string
            user:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
      tags:
      - User
      responses:
        '200':
          description: 更新成功
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  name:
                    type: string
        '400':
          description: bad request
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  mail:
                    type: string
                  name:
                    type: string
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
  /entry:
    post:
      summary: 入室処理
      parameters:
      - in: body
        name: user
        description: 入室処理
        schema:
          type: object
          required:
          - id
          - purpose
          properties:
            id:
              type: string
            purpose:
              type: string
              enum: [WORK, CIRCLE, STUDY, MEET_UP, OTHER]
      tags:
      - User
      responses:
        '201':
          description: 入室処理成功
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  mail:
                    type: string
                  isEntry:
                    type: boolean
        '400':
          description: 入室済
          schema:
            $ref: '#/definitions/ServerError'
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
  /exit:
    post:
      summary:  退出処理
      parameters:
      - in: body
        name: user
        description: 退出者情報
        schema:
          type: object
          required:
          - id
          properties:
            id:
              type: string
      tags:
      - User
      responses:
        '201':
          description: 退出処理成功
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  mail:
                    type: string
                  isEntry:
                    type: boolean
        '400':
          description: bad request
          schema:
            $ref: '#/definitions/ServerError'
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
  '/user/{id}/status':
    get:
      summary: 入室確認処理
      parameters:
      - in: path
        name: id
        type: string
        required: true
        description: 指定されたidのユーザーが入室済か返す
      tags:
      - User
      responses:
        '200':
          description: 入室済み
          schema:
            type: object
            properties:
              id:
                type: string
              user:
                type: object
                properties:
                  isEntry:
                    type: boolean
        '400':
          description: bad request
          schema:
            $ref: '#/definitions/ServerError'
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
  '/users':
    get:
      summary: 参加者一覧
      parameters:
      - in: query
        name: from
        type: string
        description: YYYYMMDD
      - in: query
        name: to
        type: string
        description: YYYYMMDD
      tags:
      - User
      responses:
        '200':
          description: 一覧取得成功
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                user:
                  type: object
                  properties:
                    name:
                      type: string
                    isEntry:
                      type: boolean
                    purpose:
                      type: string
                      enum: [WORK, CIRCLE, STUDY, MEET_UP, OTHER]
        '400':
          description: bad request
          schema:
            $ref: '#/definitions/ServerError'
        '500':
          description: サーバー側の処理に問題が発生
          schema:
            $ref: '#/definitions/ServerError'
definitions:
  ServerError:
    properties:
      error:
        type: object
        properties:
          message:
            type: string